listo los desafios para subir a github

+++ emprendedor2.rb
@@ -8,11 +8,11 @@
 
 #en caso de que la rentabilidad sea positiva, aplicar 35% impuesto
 
-price = ARGV[0].to_i
-normalusers = ARGV[1].to_i
-premiumusers = ARGV[2].to_i
-freeusers = ARGV[3].to_i
-expenses = ARGV[4].to_i
+price = ARGV[0].to_f
+normalusers = ARGV[1].to_f
+premiumusers = ARGV[2].to_f
+freeusers = ARGV[3].to_f
+expenses = ARGV[4].to_f
 
 rentability = ((price * normalusers) + ((price * 2) * premiumusers) + ((price * 0) * freeusers) - expenses)
